Add -s 'save' option and support multiple input files
The -s flag requires one argument which is a filename suffix. When enabled,
any lines removed from the input file will be appended to a file with this
suffix added to the input filename. This allows you to keep a longer period
of entries in another file (which can then be trimmed by this tool as well).

Multiple input files are now supported, which is useful because the reload
command specified with -e will only be run once for all input files. This
makes sense for, for instance, a web server which has access and error logs
being written from the same process.